Linux build works!
However, it points out an obvious flaw in what we've got going on here:
we should not be distributing binaries for GLEW and GLFW; we want to be
compiling them from scratch (with the code checked into the repo).
However, these asshole library devs seem to be a bit obsessed with VS.
So, we'll have to write our own build for them to do this. :/