Forgive me for I have sinned.
However, without a _solid_ explanation of why libgcc
is _necessary_, I can't help but feel like it's a.. eh
Lame thing about GCC? I don't know. Do other compilers
do the same thing?

I mean, I get why I need to link with libgcc. But I don't
really understand why they deemed it necessary to build
software in this way. Why is it better to link whatever
that code may or may not be as a separate library
rather than just having the compiler generate it?

*shrugs*

As far as I can tell, it doesn't interfere with our
libclessness. So. :P